backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
python-dotenv==1.0.1
openai==1.51.0
httpx==0.27.2
python-multipart==0.0.12
sse-starlette==2.1.3
aiofiles==24.1.0
